requirements.txt
numpy
pandas
scikit-learn
opencv-python
Pillow
torch
torchvision
git+https://github.com/openai/CLIP.git
ultralytics
fastapi
uvicorn[standard]
tqdm
requests
